fix: merging progress
