+++ src/main/java/ch/presentium/backend/api/schedule/mapper/SchoolClassMapper.java
@@ -19,6 +19,4 @@ public interface SchoolClassMapper {
     static String mapName(SchoolClass schoolClass) {
         return "%s-%s-%s".formatted(schoolClass.getCourse().getName(), schoolClass.getGroup(), schoolClass.getName());
     }
-
-    SchoolClass toModel(SchoolClassViewModel schoolClassViewModel);
 }
