+++ src/main/java/ch/presentium/backend/api/schedule/mapper/StudentMapper.java
@@ -1,7 +1,7 @@
-package ch.presentium.backend.api.schedule.student.mapper;
+package ch.presentium.backend.api.schedule.mapper;
 
-import ch.presentium.backend.api.schedule.student.model.StudentViewModel;
-import ch.presentium.backend.api.schedule.student.request.StudentRequest;
+import ch.presentium.backend.api.schedule.model.StudentViewModel;
+import ch.presentium.backend.api.schedule.request.StudentRequest;
 import ch.presentium.backend.business.model.user.Student;
 import org.mapstruct.Mapper;
