--- src/main/java/ch/presentium/backend/api/schedule/room/model/RoomViewModel.java
@@ -1,12 +1,0 @@
-package ch.presentium.backend.api.schedule.room.model;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class RoomViewModel {
-
-    @NotNull private final String name;
-}
