+++ src/main/java/ch/presentium/backend/business/service/StudentService.java
@@ -1,32 +1,35 @@
 package ch.presentium.backend.business.service;
 
 import ch.presentium.backend.business.model.user.Student;
-import ch.presentium.backend.business.repository.StudentRepository;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
-@RequiredArgsConstructor
-public class StudentService {
+public interface StudentService {
+    /**
+     * Find all students.
+     * @return A list of all students.
+     */
+    List<Student> findAll();
 
-    private final StudentRepository studentRepository;
+    /**
+     * Find a student by ID.
+     * @param id The ID of the student to find.
+     * @return The student with the given ID, or an empty optional if no such student exists.
+     */
+    Optional<Student> findById(UUID id);
 
-    public List<Student> findAll() {
-        return studentRepository.findAll();
-    }
+    /**
+     * Add a student.
+     * @param student The student to add.
+     */
+    void addStudent(Student student);
 
-    public Optional<Student> findById(UUID id) {
-        return studentRepository.findById(id);
-    }
-
-    public void addStudent(Student student) {
-        studentRepository.save(student);
-    }
-
-    public void deleteStudent(UUID id) {
-        studentRepository.deleteById(id);
-    }
+    /**
+     * Delete a student by ID.
+     * @param id The ID of the student to delete.
+     */
+    void deleteStudent(UUID id);
 }
