+++ src/main/java/ch/presentium/backend/business/service/StudentServiceImpl.java
@@ -1,0 +1,32 @@
+package ch.presentium.backend.business.service;
+
+import ch.presentium.backend.business.model.user.Student;
+import ch.presentium.backend.business.repository.StudentRepository;
+import java.util.List;
+import java.util.Optional;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class StudentServiceImpl implements StudentService {
+
+    private final StudentRepository studentRepository;
+
+    public List<Student> findAll() {
+        return studentRepository.findAll();
+    }
+
+    public Optional<Student> findById(UUID id) {
+        return studentRepository.findById(id);
+    }
+
+    public void addStudent(Student student) {
+        studentRepository.save(student);
+    }
+
+    public void deleteStudent(UUID id) {
+        studentRepository.deleteById(id);
+    }
+}
